ml-module/requirements.txt
GitPython==2.1.11
numpy==1.16.0
pandas==0.23.4
pika==0.13.0
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
keras
tensorflow
pillow